third_party/WebKit/LayoutTests/vr/latest/getDevices_one_device.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../resources/mock-vr-service.js"></script>
<script src="../resources/test-constants.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.vr.getDevices().then( (devices) => {
    t.step( () => {
      assert_true(devices != null);
      assert_equals(devices.length, 1);
      let device = devices[0];
      assert_equals(device.deviceName, 'Google, Inc. Daydream View');
      assert_false(device.isExternal);
    }, "getDevices returned correct results");
  }, (err) => {
    t.step( () => {
      assert_unreached("getDevices rejected");
    });
  }).then( () => {
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"navigator.vr.getDevices properly returns a single device");

</script>
